Move to workingGroupList
