+++ src/TuDa.CIMS.Web/Components/Pages/WorkingGroupListPage.razor
@@ -1,5 +1,5 @@
 @page "/working-groups"
-@using TuDa.CIMS.Web.Components.WorkingGroupPage
+@using TuDa.CIMS.Web.Components.WorkingGroupList
 
 <WorkingGroupPageWorkingGroupList CreateGroupSelected="OpenDialogAsync"/>
