+++ src/TuDa.CIMS.Web/Components/Pages/WorkingGroupListPage.razor.cs
@@ -2,21 +2,25 @@ using Microsoft.AspNetCore.Components;
 using MudBlazor;
 using TuDa.CIMS.Shared.Dtos;
 using TuDa.CIMS.Shared.Entities;
-using TuDa.CIMS.Web.Components.WorkingGroupPage;
+using TuDa.CIMS.Web.Components.WorkingGroupList;
 using TuDa.CIMS.Web.Services;
 
 namespace TuDa.CIMS.Web.Components.Pages;
 
-public partial class WorkingGroupPage
+public partial class WorkingGroupListPage
 {
     [Inject]
     private IDialogService DialogService { get; set; } = null!;
-    [Inject] private IWorkingGroupApi _workingGroupApi { get; set; } = null!;
+
+    [Inject]
+    private IWorkingGroupApi _workingGroupApi { get; set; } = null!;
+
     private async Task OpenDialogAsync()
     {
         var options = new DialogOptions { CloseOnEscapeKey = true };
         var dialog = await DialogService.ShowAsync<WorkingGroupPageWorkingGroupListAddDialog>(
-            "AddWorkingGroupDialog",options
+            "AddWorkingGroupDialog",
+            options
         );
         var result = await dialog.Result;
         var professor = await dialog.GetReturnValueAsync<Professor>();
